duplicateKeyException

+++ src/main/resources/schema-pg.sql
@@ -50,7 +50,7 @@ CREATE TABLE IF NOT EXISTS e_rst_ses_stg (
     va_nam varchar,
     dh_str timestamp(6),
     dh_end timestamp(6),
-    cd_ord smallint,
+    cd_ord int,
     cd_prn_ses UUID
 )
 PARTITION BY RANGE (dh_str);
@@ -85,7 +85,7 @@ CREATE TABLE IF NOT EXISTS e_rst_rqt_stg (
     va_nam varchar,
     dh_str timestamp(6),
     dh_end timestamp(6),
-    cd_ord smallint,
+    cd_ord int,
     cd_rst_rqt UUID
 )
 PARTITION BY RANGE (dh_str);
@@ -111,7 +111,7 @@ CREATE TABLE IF NOT EXISTS e_smtp_stg (
     dh_str timestamp(6),
     dh_end timestamp(6),
     va_cmd varchar,
-    cd_ord smallint,
+    cd_ord int,
     cd_smtp_rqt UUID
 )
 PARTITION BY RANGE (dh_str);
@@ -150,7 +150,7 @@ CREATE TABLE IF NOT EXISTS e_ftp_stg (
     dh_end timestamp(6),
     va_cmd varchar,
     va_arg varchar,
-    cd_ord smallint,
+    cd_ord int,
     cd_ftp_rqt UUID -- index
 )
 PARTITION BY RANGE (dh_str);
@@ -177,7 +177,7 @@ CREATE TABLE IF NOT EXISTS e_ldap_stg (
     dh_end timestamp(6),
     va_cmd varchar,
     va_arg varchar,
-    cd_ord smallint,
+    cd_ord int,
     cd_ldap_rqt UUID
 )
 PARTITION BY RANGE (dh_str);
@@ -210,7 +210,7 @@ CREATE TABLE IF NOT EXISTS e_dtb_stg (
     va_cnt varchar,
     va_cmd varchar,
     va_arg varchar,
-    cd_ord smallint,
+    cd_ord int,
     cd_dtb_rqt UUID
 )
 PARTITION BY RANGE (dh_str);
@@ -235,7 +235,7 @@ CREATE TABLE IF NOT EXISTS e_exc_inf (
     va_err_typ varchar,
     va_err_msg varchar,
     va_stk json,
-    cd_ord smallint,
+    cd_ord int,
     cd_rqt UUID
 );
