+++ src/main/resources/schema.sql
@@ -48,7 +48,7 @@ CREATE TABLE IF NOT EXISTS e_rst_ses_stg (
     va_nam varchar,
     dh_str timestamp(6),
     dh_end timestamp(6),
-    cd_ord smallint,
+    cd_ord int,
     cd_prn_ses UUID
 );
 
@@ -81,7 +81,7 @@ CREATE TABLE IF NOT EXISTS e_rst_rqt_stg (
     va_nam varchar,
     dh_str timestamp(6),
     dh_end timestamp(6),
-    cd_ord smallint,
+    cd_ord int,
     cd_rst_rqt UUID
 );
 
@@ -105,7 +105,7 @@ CREATE TABLE IF NOT EXISTS e_smtp_stg (
     dh_str timestamp(6),
     dh_end timestamp(6),
     va_cmd varchar,
-    cd_ord smallint,
+    cd_ord int,
     cd_smtp_rqt UUID
 );
 
@@ -142,7 +142,7 @@ CREATE TABLE IF NOT EXISTS e_ftp_stg (
     dh_end timestamp(6),
     va_arg varchar,
     va_cmd varchar,
-    cd_ord smallint,
+    cd_ord int,
     cd_ftp_rqt UUID -- index
 );
 
@@ -167,7 +167,7 @@ CREATE TABLE IF NOT EXISTS e_ldap_stg (
     dh_end timestamp(6),
     va_arg varchar,
     va_cmd varchar,
-    cd_ord smallint,
+    cd_ord int,
     cd_ldap_rqt UUID
 );
 
@@ -198,7 +198,7 @@ CREATE TABLE IF NOT EXISTS e_dtb_stg (
     va_cnt varchar,
     va_arg varchar,
     va_cmd varchar,
-    cd_ord smallint,
+    cd_ord int,
     cd_dtb_rqt UUID
 );
 
@@ -221,7 +221,7 @@ CREATE TABLE IF NOT EXISTS e_exc_inf (
     va_err_typ varchar,
     va_err_msg varchar,
     va_stk text,
-    cd_ord smallint,
+    cd_ord int,
     cd_rqt UUID
 );
